requirements.txt
pyspark>=3.3.0
delta-spark>=2.1.0
pandas>=1.3.0
seaborn>=0.11.0
matplotlib>=3.4.0
